# Add AVX optimizations for pdist