--- a/aten/src/ATen/cpu/vec256/functional.h
+++ b/aten/src/ATen/cpu/vec256/functional.h
@@ -68,8 +68,8 @@ template <typename scalar_t, typename MapOp, typename ReduceOp>
 inline scalar_t map2_reduce_all(
     const MapOp& map_fun,
     const ReduceOp& red_fun,
-    scalar_t* data,
-    scalar_t* data2,
+    const scalar_t* data,
+    const scalar_t* data2,
     int64_t size) {
   using Vec = vec256::Vec256<scalar_t>;
   if (size < Vec::size) {