@erikbrinkman 2018-09-05
In general I've just been using ipythons `%timeit` to run the command at least 10 times and report the timing.

The script for the comparison to scipy looks something like:

```python
import torch
import scipy.spatial.distance as spd

xt = torch.randn(2048, 2048)
xn = xt.numpy()

for p in [0, 1, 2, 3, float('inf')]:
    if p == 0:
        spdist = lambda x: spd.pdist(x, 'hamming') * x.shape[1]
    elif p == float('inf'):
        spdist = lambda x: spd.pdist(x, lambda row, col: np.abs(row - col).max())
    else:
        spdist = lambda x: spd.pdist(x, 'minkowski', p=p)

    print(p)
    %timeit torch.pdist(xt, p)
    %timeit spdist(xn)
```