@cpuhrsch 2018-09-05
@erikbrinkman  - In general it should be worth running this on a few different magnitudes of tensors (so 10^3/4/5 elements) to see how it scales.