@ezyang 2018-09-05
Can https://pytorch.org/docs/stable/nn.html?highlight=pairwise#torch.nn.functional.pairwise_distance benefit from this vectorized code now?