@erikbrinkman 2018-09-06
@cpuhrsch The default dispatch is still slower than it was before introducing this change. The cause isn't clear other than it seems that the compiler is not eliding all of the Vec256 operations. If the answer is that lack of AVX is an uncommon use case and we don't care about performance there than this seems ready to push. If the answer is that we still want this to be _reasonably_ fast than there are two ways to fix it: making the dispatch dispatch to the old code, or introducing tricks to get the compiler to figure out how to handle vec256_base functions appropriately. The later seems far outside the scope of this PR.

In response to timing for various sizes single threaded, usually an average of 7 runs. These two graphs are with the other dimension set to 2048. Note that for changing the number of rows, the computation grows by n^2. The width of each bar corresponds to the uncertainty. The scaling on each seems appropriate.

![num_cols](https://user-images.githubusercontent.com/858926/45127838-2fe64700-b12f-11e8-8d52-e086a0b474eb.png)

![num_rows](https://user-images.githubusercontent.com/858926/45127843-34aafb00-b12f-11e8-936b-981114af2a5f.png)